Embedded spotify player, link parsing

+++ sql/follow.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `collective`.`follow` (
-  `user_id` int(11) unsigned NOT NULL,
-  `follow_user_id` int(11) unsigned NOT NULL,
+  `user_id` bigint(24) unsigned NOT NULL,
+  `follow_user_id` bigint(24) unsigned NOT NULL,
   `last_update` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   `deleted` tinyint(1) NOT NULL DEFAULT '0',
   PRIMARY KEY (`user_id`, `follow_user_id`)
