+++ sql/status.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `collective`.`status` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_id` int(11) unsigned NOT NULL,
+  `id` bigint(24) NOT NULL AUTO_INCREMENT,
+  `user_id` bigint(24) unsigned NOT NULL,
   `msg` varchar(255) NOT NULL,
   `link` varchar(255) NOT NULL,
   `date` timestamp NOT NULL DEFAULT '1970-01-01 00:00:01',
